+++ src/store/reducer.js
@@ -1,22 +1,7 @@
 const initialState = {
-  // tasks:  [
-  //   {
-  //     tag: "🧹 chore",
-  //     taskName: "Water plant",
-  //     taskDescription: "",
-  //     dueDate: "today",
-  //     hashKey: 1184996857
-  //   },
-  //   {
-  //     tag: "💼 work",
-  //     taskName: "Udemy",
-  //     taskDescription: "React course. At least 5 lectures!",
-  //     dueDate: "",
-  //     hashKey: 1666272339
-  //   }
-  // ],
+  networkError: false,
+  token: null,
   tasks: [],
-  networkError: false
   // tags: [] // eventually
 }
